requirements.txt
appdirs==1.4.4
argcomplete==3.1.4
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
cattrs==22.2.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.1.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.46.0
fs==2.4.16
fuse-python==1.0.7
gpg==1.18.0
gyp==0.1
html5lib==1.1
httplib2==0.20.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libvirt-python==10.0.0
louis==3.29.0
lsprotocol==2023.0.0
lxml==5.2.1
lz4==4.0.2+dfsg
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
mechanize==0.4.9
mpmath==0.0.0
mutagen==1.46.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
packaging==24.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
psutil==5.9.8
ptyprocess==0.7.0
pyasyncore==1.0.2
pycairo==1.25.1
pycryptodomex==3.20.0
pycuda==2024.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.7.0
pylibacl==0.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytools==2023.1.1
pytz==2024.1
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
ruffus==2.8.4
SciPy==1.11.4
screen-resolution-extra==0.0.0
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
ssh-import-id==5.11
sympy==1.12
systemd-python==235
tornado==6.4
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
wadllib==1.3.6
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.1
wheel==0.42.0
xdg==5
xkit==0.0.0
yt-dlp==2024.4.9
